Remove last explainTypeMismatch special case (#8598)
* Stop treating MetaType as a proxy type

There's a lot of code checking `is_proxy_type` assuming that it's going
to find a literal type. That's not really true, and also the
`underlying()` method of a `MetaType` returns `Object` which is like
aspirationally true but not really useful and if we're depending on
this, it's probably a bug in disguise.

* Move out the last explainTypeMismatch special case

* Remove explainTypeMismatch

* stog //test/cli:update_runtime_types

* format

* Capture changes to the tests

* Still have to treat it like their underlying is Object

There are 41 type errors in Stripe's codebase from this and then a
handful more from typechecking sorbet-runtime, and I don't actually want
to fight this battle today: I just want to deprecate this
`expliainTypeMismatch` method.

Most of the violations come from:

- `==`
- `!=`
- `puts`

So it's reasonable to assume that we could do this migration in the
future and stop somtimes treating metatypes like `Object` but that's a
battle for another day.

* put these tests back