Propagate packager options when copying GlobalState
The core problem here is that when we would hash a file in isolation (in
hashing.cc), we would not also run packager on the file: we would only
run `indexOne` and `nameAndResolve`.

At this point, there are only three usages of `nameAndResolve`:

1.  The call from `populate.cc`, where we make the binary payload.
2.  The call from `fuzz_dash_e` where we attempt to recreate Sorbet's
    pipeline with as little overhead as possible for the sake of
    fuzzing. Probably completely broken at this point, because it's not
    used.
3.  The call from `hashing.cc`.

We don't currently read any of the options out of `packageDB` in either
namer or resolver, but we are going to start doing that as we move more
of the packager out of `packager.cc` and into Sorbet's core pipeline. In
that case, it's important that the `enabled_` boolean on the `packageDB`
be accurate.

This boolean is currently only used in one place: the `toProto` logic
when serializing the file table, so the fact that it didn't do the right
thing was not really load bearing yet, and thus cannot be tested.