VS Code is sometimes sending `source.fixAll` (#8642)
* Add failing test

* VS Code is sometimes sending `source.fixAll`

... despite the fact that we never declare that we will send it.
I'm not sure if this represents a misconfiguration on the user's part, a
bug in VS Code, a faulty implementation of Sorbet in the LSP spec, or
what.

But what's definitely Sorbet's fault is that code action kind is an open
enum, and Sorbet should not fail for unknown values.

That will take some time to fix (but is fixable in the future). In the
mean time, to unblock users, I'm just going to allow that code action
kind.

cc @damolina-stripe

* Fix failing test