Delete leading whitespace from duplicate imports
This isn't _amazing_ when there is a trailing comment on the current and
previous line, because they will get combined.

Maybe a better option would be to leave the leading whitespace if the
line would not be empty as a result of removing the import, but even
that wouldn't be perfect, because Stripe's other `__package.rb` parser
requires that comments must be exactly two space indented.

So for now I think this is the easiest-to-implement solution to the
problem of Sorbet autocorrects leaving the file in a malformed state.