Make `isPackage` check if the packager is enabled
This should prevent a class of error where people ask "is this a
`__package.rb` file?" and then do something where they assume an answer
of "yes" also implies that `--stripe-packages` was passed.

Note that this makes it easier to accidentally start depending on
whether the `--stripe-packages` option was set in places where we
shouldn't, but I think we should address that in a future change, by
invalidating the cache if certain known-problematic command line options
change.