Make `isPackage` check if the packager is enabled (#8635)
* isPackageRBI → hasPackageRBIPath

* Make `isPackage` check if the packager is enabled

This should prevent a class of error where people ask "is this a
`__package.rb` file?" and then do something where they assume an answer
of "yes" also implies that `--stripe-packages` was passed.

Note that this makes it easier to accidentally start depending on
whether the `--stripe-packages` option was set in places where we
shouldn't, but I think we should address that in a future change, by
invalidating the cache if certain known-problematic command line options
change.

* Thread through a GlobalState argument

Some interesting callouts:

- I've kept certain now-redundant checks to `stripePackages`, for
  performance (short circuiting in loops).
- There is one raw read to `hasPackageRbPath` in `runAutogen` because I
  wasn't sure whether it was load bearing or not (e.g., maybe Stripe is
  running autogen without `--stripe-packages` but still want to treat
  `__package.rb` files as not real files).

Technically I could write tests for features like the references or
workspace_symbols to assert that they start working again in
`__package.rb` files, but I don't really think that's super worth it. If
you think so, let me know. (Same for the resolver changes.)

* Fix tests

* Be sure to always set the packager options