Delete --package-rbi-generation mode
The `--package-rbi-generation` mode is proving difficult to maintain for
a couple reasons:

- It initializes the packager options differently from the rest of the
  pipeline. This is important, because we need the GlobalState to have
  the right packager options so that `namer` can decide whether it wants
  to treat `__package.rb` files as packages and enter them into the
  symbol table, or leave them as normal Ruby files.

- It does weird things with `__package.rb` files. It tries to make it
  seem like you're running with `opts.stripePackages` being false, but
  then do package-related things (e.g., `findPackages`).

- It's going to complicate our attempts to interleave resolve and
  typecheck.

This feature never made it into use at Stripe, and never gained LSP
support. Also, we generally don't think this approach is the approach
we'd use if we were to revisit that: we think that we'd probably have a
"one RBI per package" approach, instead of "one RBI per transitive
package deps" approach, which would get rid a lot of the hacks around
"this symbol doesn't resolve" that have infected resolver right now.

This change deletes all but the code in `packager/rbi_gen.{h,cc}`. The
RBI generation code itself is useful to keep around because it has a
bunch of stuff which could be useful for simply generating RBIs from
`GlobalState`.