Remove test-rbi-gen.sh build job
The code in `packager/rbi_gen.*` will still be tested indirectly via the
`//packager:packager` library.

We could split it out, but then we'd have to make a whole job just to
test it and there's not that much code in the library in the first
place, so I'm not too worried about it being separate.