Create a minimize GlobalState from scratch
We used to use `deepCopy` as a way to avoid needing to duplicate all the
option parsing.

Now that the option parsing is in its own helper, we can just call that.

This has the added benefit that the `GlobalState` we get is even more
empty: it doesn't have any of the things in the kvstore. Another benefit
is that all the `minimizeRBI` code lives in one place, instead of being
sprinkled throughout the function.