Revert "Add an ENFORCE"
This reverts commit 7183e5a7b896424c120d13dbc5999c280d3e26bb.

We can't add this seemingly sensible ENFORCE because the rbi-gen code
plays fast and loose with the pipeline. Those two tests declare
`# enable-packager: true`, in spite of the normal pipeline saying that
`--stripe-packages` and `--package-rbi-generation` are incompatible.

I'd rather not think about it, because the more I look at the rbi-gen
code the less I like it and the more I want to remove it. I'm just going
to re-remove the ENFORCE, because I only added it speculatively—it
doesn't accurately describe the state of the world unless we remove
rbi-gen mode.