Move setPackagerOptions into setGlobalStateOptions (#8671)
* Create a minimize GlobalState from scratch

We used to use `deepCopy` as a way to avoid needing to duplicate all the
option parsing.

Now that the option parsing is in its own helper, we can just call that.

This has the added benefit that the `GlobalState` we get is even more
empty: it doesn't have any of the things in the kvstore. Another benefit
is that all the `minimizeRBI` code lives in one place, instead of being
sprinkled throughout the function.

* Move setPackagerOptions into setGlobalStateOptions

* Inline pipeline::setPackagerOptions into setGlobalStateOptions

There should be no places where we care about command-line options and
do not also care about correctly setting up the packager options.

* Actually copy everything in PackageDB::deepCopy

We were missing some things, which were getting hidden by the fact that
we called `pipeline::setPackagerOptions` again in LSPTypechecker.cc
(right before the call to `pipeline::package`).

Once we removed that, it became clear that we were missing these
options.

* Remove calls to setPackagerOptions in pkg_autocorrects_test

* Add an ENFORCE

* Add some comments

* Revert "Add an ENFORCE"

This reverts commit 7183e5a7b896424c120d13dbc5999c280d3e26bb.

We can't add this seemingly sensible ENFORCE because the rbi-gen code
plays fast and loose with the pipeline. Those two tests declare
`# enable-packager: true`, in spite of the normal pipeline saying that
`--stripe-packages` and `--package-rbi-generation` are incompatible.

I'd rather not think about it, because the more I look at the rbi-gen
code the less I like it and the more I want to remove it. I'm just going
to re-remove the ENFORCE, because I only added it speculatively—it
doesn't accurately describe the state of the world unless we remove
rbi-gen mode.