Avoid possibly using a freed global state from a worker (#8679)
There's a potential race condition where one worker finishes
substitution for all trees before another starts. In this case, it's
possible for the indexer to be canceled, and the global state for the
slow path to be freed before the substituteTreesWorker is started.

As the spdlog instance will be long-lived (it belongs to an error queue
that lives for the duration of the LSPTypechecker) taking a reference to
that to use for tracing avoids a potential use-after-free of `cgs`.