Prevent creating files larger than 2^31 bytes (#8656)
* Upgrade INVALID_POS_LOC to UINT32_MAX

Right now, it's `0xfff'ffff`. Note that it's NOT:

- `0x7fff'ffff`
- `0xffff'ffff`

which would have maybe been sensible. Instead, it's currently `2^28-1`,
despite the `beginPos` and `endPos` fields both holding `uint32_t`
(i.e., pretending that they have `2^32` bits available).

* Prevent creating files larger than 2^32 bytes

* Use `.exists()` in Loc.h

This is a mess.

`INVALID_POS_LOC` is defined as `UINT32_MAX`. When we ask whether a
`uint32_t` is less than or equal to the max `uint32_t`, this is of
course always true.

So in effect, these `ENFORCE` are not testing anything.

What we probably want to test instead is that when you create a loc from
two offsets directly, those correspond to valid positions. The
definition of `LocOffsets::exists` checks that the `begin`/`end` offset
are specifically not equal to `INVALID_POS_LOC` (e.g. strictly less
than).

The problem is that any time we call the default (zero-argument)
constructor of a `Loc`, we call it with `LocOffsets::none()` which by
default is a `LocOffsets` that does not exist, so we can only check that
the `LocOffsets` exists when explicitly given a `begin` and `end`.

By defining two independent constructors, we can have the `.exists`
logic only apply if someone constructs a `Loc` from explicit `begin` and
`end` pos, not from an already-initialized `LocOffsets::none()`.

* Move `begin <= end` ENFORCE to LocOffsets.h

We don't do this logic in teh default constructor because we don't need
to, `INVALID_POS_LOC` is of course equal to itself.

* Trevor's curly bracket suggestion