Store lineBreaks with uint32_t (#8678)
* Add a comment about the c_lower_bound thing

* Store lineBreaks with uint32_t

This lets us not lie when we call `c_lower_bound`, because we need the
list be sorted, but it appears not sorted when the comparison forces an
upcast to unsigned values (and the `-1` became `UINT32_MAX`).

* Fix ENFORCE

There is one fewer element in `lineBreaks.size()` now, so the ENFORCE
also needs adjusting.

* Add another comment

* Apply suggestions from code review

Co-authored-by: Trevor Elliott <awesomelyawesome@gmail.com>

---------

Co-authored-by: Trevor Elliott <awesomelyawesome@gmail.com>