Delete leading whitespace from duplicate imports (#8646)
* prework: Add extra test: `test_import`

(We already have one for `import`)

* prework: Convert to `^^^` format for "Duplicate" assertions

* Delete leading whitespace from duplicate imports

This isn't _amazing_ when there is a trailing comment on the current and
previous line, because it won't be included in the portion that's
deleted.

We could delete the whole line that the import is on, but in practice,
we have other tools that enforce that there are no trailing line
comments in package files.