`--max-threads=0` to make a test not flake
This test has flaked twice in the past few days:

- <https://buildkite.com/sorbet/sorbet/builds/35161#0195edf0-655c-4b38-8bd5-eb784c5f7f02/140-376>
- <https://buildkite.com/sorbet/sorbet/builds/35147#0195e053-8775-46aa-a905-431dc2d7f7a2/213-462>

We use `--max-threads=0` in ~all of our other `--stripe-packages` tests,
because they tend to have multiple files, which might get scheduled onto
workers in different orders every run.

Spot checking the failures, it seems like that's the only problem with
the flakes this time (that sometimes the errors for one file show up
sooner or later in the output).