Interleave packager and namer calls in LSPTypechecker (#8745)
* prework: Remove redundant stripePackages check

If stripePackages is not set, then `indexed` will just be an empty
vector and thus this call is a no-op.

* clang-format

* Move packager after deepCopy

* clang-format

* Revert "Fix the packageDB persistence bug"

Reverts commit 07e0e5fdae59064e93f72d7355c75d4060bf470c

This doesn't actually break things, because by moving the calls to
packager below, we fixed the bug in another way.

* Interleave packager and namer calls

This is more accurate w.r.t. how realmain does it.