Move some PackageSpec rewrites to `rewriter` (#8727)
* Move stripePackages to CacheSensitiveOptions

* Change call site to read cacheSensitiveOptions

* Add rewriter/PackageSpec scaffolding

* Move some functions from packager to rewriter

* Remove outdated comment

* Use `beginIndexerError` to report these errors

Must use `MutableContext` because `beginIndexerError` also mutates the
`core::File` object on which the error is reported (to record that there
were indexer errors).

* prework: Move populateMangledName down

It's only used in `createAndPopulatePackageInfo`,, it should probably
live closer to down there.

* Use super class to find the PackageSpec ClassDef

This invariant has been established by rewriter, so it's available for
us to depend on here.

This means that we can delete `startsWithPackageSpecRegistry`, a
recursive function that would walk up a constant lit to see whether it
started with `<PackageSpecRegistry>` to detect whether a class is a
PackageSpec.

* Document PackageSpec.h

* Inline a helper function

* fixup! Change call site to read cacheSensitiveOptions

* Actually enable the packager when indexing

* Make pkg_autocorrects_test.cc run rewriter