getFiles: `continue;` → `subspan(1)`
In 5865c505bdb30e80b1a1861a92b6a62d751d8827, Trevor made it so that
`getFiles` returns a span instead of a const vector, which means that we
now have access to the `subspan` method, and we can avoid doing some
branches inside a loop.

Slight performance improvement, but really just a nice to have.