Use `MangledName` in `PackageInfo` interface
We're trying to move away from identifying packages by `vector<NameRef>`
and move them to be their own symbol kind.

This does not yet touch exports: exports are constants, not packages,
and changing the data model so that we have a `vector<SymbolRef>` for
exports instead of `vector<vector<NameRef>>` will be a larger change.