Rework `visibilityApplies` to work on `MangledName`
This is a bit clunky: we were `absl::Span` equality to compute wildcard
visibility. To do that with mangled names, we have to use string
`StartsWith` checking, plus a check for boundary conditions.

This check currently relies on the fact that package names are not
allowed to have `_` in them (this is checked by Sorbet).

In the future, when we get rid of `MangledName` entirely in favor of
package symbols, we can reimplement this check by following the
`->owner` chain on the symbols to see if we have one package that is
owned by another package, which will mean we no longer assume
`_`-free package names.