Get rid of `_Package` suffix from `MangledName` (#8751)
We don't actually need this `_Package` suffix.

I'm working on some changes which will completely replace `MangledName`
with `PackageRef` IDs. Eventually, nothing will operate on
`MangledName`.

I also thought that we might have an optimization that meant that the
strings backing two names like `Opus_Foo` and `Opus_Foo_Bar` could
overlap in the string table, but I don't actually know if that's the
case now that I'm looking at the code—it looks like if a name doesn't
exist we unconditionally copy the whole string into the string table.

Regardless, this should save memory because we're not storing `8 *
num_packages` extra bytes in the name table.