packager.cc: `prependRoot` is obsolete (#8763)
* packager.cc: `prependRoot` is obsolete

Previously, we needed to prepend `Symbols::root()` to every `export`
because the rewrite-based approach to package visibility meant that the
exported constants would not actually resolve the right way.

Now that we have dramatically simpler visibility checking, we can stop
doing these rewrites. The exported constant symbols and the package
declarations are always in different namespaces, so it's as if they're
always resolving from `Symbols::root()`.

* tools/scripts/update_testdata_exp.sh

* stog   //test/cli:update_package-export-suggestion-boundary \