Use `MangledName` in `PackageInfo` interface (#8757)
* Use `MangledName` in `PackageInfo` interface

We're trying to move away from identifying packages by `vector<NameRef>`
and move them to be their own symbol kind.

This does not yet touch exports: exports are constants, not packages,
and changing the data model so that we have a `vector<SymbolRef>` for
exports instead of `vector<vector<NameRef>>` will be a larger change.

* Trivial updates, no explanation needed

* Rework `visibilityApplies` to work on `MangledName`

This is a bit clunky: we were `absl::Span` equality to compute wildcard
visibility. To do that with mangled names, we have to use string
`StartsWith` checking, plus a check for boundary conditions.

This check currently relies on the fact that package names are not
allowed to have `_` in them (this is checked by Sorbet).

In the future, when we get rid of `MangledName` entirely in favor of
package symbols, we can reimplement this check by following the
`->owner` chain on the symbols to see if we have one package that is
owned by another package, which will mean we no longer assume
`_`-free package names.

* Change information reported by `dumpPackageInfo`

The `--dump-package-info` flag is only used in tests as a sanity check.

Its output is not used outside of Sorbet, and the contents are never
read. We can choose whatever debug information we want in this file.

* One more place we don't want fullName

* Revert changes to rbi_gen runner

* Add test for shared prefixes