Remove dead code from resolver (#8769)
This code is dead. The commit that added the comment added a test
exactly like what the comment looks like, and the test passes even after
this logic is deleted.

This code is dead because `T.let` will get translated into a `Cast`
node, and then not hit this codepath anymore (e.g., there will not be a
`send` node at all).