Stop relying on `ObjectSpace._id2ref`
This method was supposed to be deprecated for Ruby 2.7 but fell
through the cracks, I'm working on making it happen:
https://github.com/ruby/ruby/pull/13157

In addition, Ruby 3.5.0 already has some optimizations in place that
are de-optimized if `_id2ref` is ever called:
https://github.com/ruby/ruby/pull/13115

And one more issue is that `_id2ref` doesn't work well with ractors.

Instead we can hold our own weakmap. It's probably not the ideal
implementation, but I chose to not go for too large of a refactor
given my limited familiarity with sorbet.