Allow passing `*.exp` paths to update_testdata_exp.sh
Sometimes I want to be able to regenerate all the `*.exp` files for a
certain kind of expectation. There are two ways to do this:

1.  regenerate everything
2.  list out each `*.rb` file that has that kind of exp

Doing the second approach always requires writing some fiddly shell
code, and it's easier to just have it here. Now you can do something
like this:

    tools/scripts/update_testdata_exp.sh test/testdata/**/*.package.exp

which saves time.