Make EnforcePackagePrefix operate on `ConstantLit`
Since packager runs after namer, all the `ClassDef` and `Assign` nodes
will actually have resolved `ConstantLit`.

In theory that should make it possible to delete a bunch of code which
implements `EnforcePackagePrefix` right now--that whole portion of
`packager.cc` operates on `MangledName` and `vector<NameRef>`, partly
because since it was previously operating on `UnresolvedConstantLit`,
there was no better way to do things.

Unfortunately, making this phase of Sorbet work entirely on symbols
instead of on `NameRef` and `MangledName` is still a little annoying.
I think it will be extremely easy once packages are in the symbol table,
so I'm deferring that work for now, and focusing on only the minimal
work required to get the namer and packager phases flipped.