Flip pipeline::name and pipeline::package (#8753)
* Flip pipeline::name and pipeline::package

* The ConstantLit in the ClassDef is resolved now

* Handle ConstantLit in getPackageName

* sbg && tools/scripts/update_testdata_exp.sh --no-build test/testdata/**/*.package-tree.exp

There are two classes of changes here:

- `UnresolvedConstantLit` nodes become `ConstantLit` nodes

  When this happens, we stop printing the `<C Foo>` names for the
  unresolved scopes and instead print the `name` field of symbol (with
  the `original->name` at the very end)

- `MethodDef` nodes have their `symbol` hidden: we would only print the
  `<<todo method>>` when the `symbol` on the `MethodDef` is different
  from the `name` in the tree. Since the methods are now resolved, and
  their name is not `<todo method>`, that part drops off.

* Make EnforcePackagePrefix operate on `ConstantLit`

Since packager runs after namer, all the `ClassDef` and `Assign` nodes
will actually have resolved `ConstantLit`.

In theory that should make it possible to delete a bunch of code which
implements `EnforcePackagePrefix` right now--that whole portion of
`packager.cc` operates on `MangledName` and `vector<NameRef>`, partly
because since it was previously operating on `UnresolvedConstantLit`,
there was no better way to do things.

Unfortunately, making this phase of Sorbet work entirely on symbols
instead of on `NameRef` and `MangledName` is still a little annoying.
I think it will be extremely easy once packages are in the symbol table,
so I'm deferring that work for now, and focusing on only the minimal
work required to get the namer and packager phases flipped.

* Two more fixes