Avoid wrapping on RBS quick reference table
The `type_parameters` example is long and causes line wrapping of
itself, and other code snippets.

We can avoid this by switching to a multiline code block for the Sorbet
column of this table (which also has the benefit of allowing us to
highlight Ruby syntax).

This comes at the cost of needing to switch to an HTML table (interlaced
with Markdown) which is harder to edit.  However, this tradeoff is
acceptable, as we should optimize for documentation readers, not for
documentation authors (and LLMs can help edit the docs anyway).