Use a `Set` for `T.deprecated_enum` (#8837)
* prework: Add test showing current behavior that should fail

* Use a `Set` for `T.deprecated_enum`

This is a performance optimization. This is technically a backwards
incompatible change, but this feature is deprecated, so I think that's
reasonable. Most things will stay the same, e.g. if you're only using
`Enumerable` APIs over the `.values` of a `T::Types::Enum` object,
you won't even notice (except performance of some operations).

* People are relying on the Hash behavior

* Make test fail and change it after fixing the code

* Update tests that change

* This is a dumb rule