Process unchanged __package.rb files on the fast path (#8838)
* Return the old File when replacing a file

This delays destruction of the shared_ptr to the caller, but doesn't
modify the reference count.

* Process __package.rb files on the fast path

We already guard __package.rb file changes from the fast path, but if
the file hasn't changed allow it to go through. However, we would also
filter __package.rb files out of the updated file set before running the
pipeline, which would lead to running the fast path on zero files.
Instead, let's build up the package files in a separate vector that we
extend with related package files, and merge right before the pipeline.

* Some tests no longer work on the fast path

* No need to force the incremental namer for __package.rb edits

* Track hashes for __package.rb files as well

* Move self-import checking to `validatePackage`