Fix NUMBER_OF_FLAGS 7 → 6
Between when the MacBook died and when the new Mac Mini was brought
online, ee56bd1532 landed, which did not get tested on macOS because the
machine was offline.

This change has a bug--it removed a cache-sensitive option but did not
also drop the NUMBER_OF_FLAGS.

For some reason, only the macOS sanitizers were powerful(?) enough to
catch this. On macOS, the sanitizers catch this failure 100% of the
time. On Linux (x86_64 was all I tested), it made this test flake with
50% probability.