bestEffortCommit the cache after writing the VERSION_KEY (#8878)
This is just a slight performance optimization. When the cache is cold,
we would write the VERSION_KEY but forget to commit the transaction.
This meant that the first write to the cache when processing files would
defensively abort any previous transaction, causing the VERSION_KEY to
get written again.

It's not clear that this should be user visible, and it's really not a
big performance win, but it looks nicer in the logs because we see
write, commit, write, commit, write, commit, instead of write, drop,
write, commit, write, commit.