Handle proxy types in type_definition.cc (#8897)
* Handle proxy types in type_definition.cc

I only tested this in my editor.

Getting an automated test for it was hard, because our `type-def`/`type`
assertions don't have a way to assert that the place you get jumped to
is someplace in the standard library, but I confirmed that's what
happens in the editor.

* Remove a test that starts failing

This doesn't go to `(nothing)` anymore--it goes to `symbol.rbi` in the
payload.

As per PR description, we don't have a way to capture that in the test
harness so I'm just deleting.