Remove `restrict_to_service` from PackageSpec
Sorbet is not actually using this annotation anything. The only thing
that this checked in practice was that the name written here resolved to
a package name (vs being some random class or module name).

But this annotation is so rarely used in Stripe's codebase (and it's not
clear whether or not we just want to remove it entirely) that it doesn't
seem worth having special logic in packager.cc just to do that rewrite.