prework: Use LocOffset throughout ParsedSig
The only place where we actually might have made an argument in favor of
keeping this as a `core::Loc` is that we actually treat `ParsedSig` as
an `in` param when we call `TypeSyntax::getResultType` in `resolver.cc`
(so that we can get the right type for parsing `T.cast` assertions that
have `T.type_parameter(:U)` things in them).

But we only use that for the `type` property of the `typeArgSpec` right
now--we don't read the `loc` as an `in` param. So it's ~fine if the Loc
on the TypeParameter is not in the same file as the `T.cast` that we're
parsing right now (which it might not be if this method is defined in
say an RBI file).