Remove origSend from ParsedSig (#8981)
* Remove origSend from ParsedSig

This `const_cast` was not great.

I think that a previous version of Sorbet (possibly something related to
the compiler) would use the `origSend` to mutate the tree. That's not
happening anymore: all call sites accessing the `origSend` is not
mutating the tree.

Also, the main user of `origSend` used to be resolver, but resolver
doesn't use this anymore. Now, it's only used indirectly by `SigFinder`.
To fix that, I'm going to have `SigFinder` return the send, and also
make it return something that's `const` (because it already takes a
`const` tree as an argument).

* Move `origSend` to the `SigFinder`

* This is the effect of making it const

With more rigorous const correctness, the compiler flags that the
`block` variable here cannot be declared as a pointer to a mutable
`ast::Block`.

* Fix sanitizer errors

* Make it be a `const &`