RBS: Fix comment association inside setter sends
When passing a send to a setter call, we were not descending inside the
send node thus not associating the RBS comment found inside.

This:

```rb
x.foo = puts(
  *[], #: untyped
)
```

Is now properly translated as:

```rb
x.foo = puts(
  *T.unsafe([]),
)
```

We can also simplify the cases for a bracket assign and a setter assign as
they are handled the same.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>