Rewrite error message with anonymous splat params (#8997)
* add test exercising confusing error message

Signed-off-by: Cameron Wong <cwong@stripe.com>

* add kwarg splat test

Signed-off-by: Cameron Wong <cam@camdar.io>

* fix confusing error with anonymous splats

I'm not in love with this phrasing (initial draft had something along
the lines of `<anon>`), but it's consistent with how the argument is
referred to in the "Malformed `sig`" message.

Signed-off-by: Cameron Wong <cam@camdar.io>

* add test for anon block args

Signed-off-by: Cameron Wong <cam@camdar.io>

* purge `isAnonymous`

After speaking with jez offline, we concluded that there really isn't a
great way to implement `isAnonymous` that doesn't require one of

a) changing how argument name normalization works
b) rendering the argument name

Option (a) seems overkill for such a tiny edge case, but hiding a call
to `argumentName` behind an innocent name like `isAnonymous` is also a
bad idea.

Signed-off-by: Cameron Wong <cam@camdar.io>

* style fixes

Signed-off-by: Cameron Wong <cam@camdar.io>

---------

Signed-off-by: Cameron Wong <cwong@stripe.com>
Signed-off-by: Cameron Wong <cam@camdar.io>