Allow customizing the untyped code highlight diagnostic severity (#9050)
* prework: enableHighlightUntyped -> initialEnableHighlightUntyped

This is never modified after the fact, because the
LSPClientConfiguration is `const`.

* Allow customizing the untyped code highlight diagnostic severity

We have been getting complaints that using information diagnostics to
highlight untyped code is too noisy sometimes especially because it
pollutes the `Problems` tab.

VS Code treats the `Hint` diagnostic level as something that never shows
up in the problems tab, and this might be a better option for some
users.

It would be nice if VS Code made it possible to easily style hint
diagnostics to have the same squiggle underline that other diagnostics
have, but that does not seem to be something VS Code allows (we might
build custom, opt-in logic in the VS Code extension to allow styling the
untyped code hint diagnostics, but that can be a future change).