Remove pakcage-special-allowed-dsls
I could leave this test around, but it would just be a bunch of "method
does not exist" errors and we already test that elsewhere.

At this point, we've removed both `autoloader_compatibility` and
`restrict_to_service`, which were the only `__package.rb` DSLs that
required special handling inside Sorbet.