Remove `restrict_to_service` from PackageSpec (#8942)
* Remove `restrict_to_service` from PackageSpec

Sorbet is not actually using this annotation anything. The only thing
that this checked in practice was that the name written here resolved to
a package name (vs being some random class or module name).

But this annotation is so rarely used in Stripe's codebase (and it's not
clear whether or not we just want to remove it entirely) that it doesn't
seem worth having special logic in packager.cc just to do that rewrite.

* Remove pakcage-special-allowed-dsls

I could leave this test around, but it would just be a bunch of "method
does not exist" errors and we already test that elsewhere.

At this point, we've removed both `autoloader_compatibility` and
`restrict_to_service`, which were the only `__package.rb` DSLs that
required special handling inside Sorbet.