Factor helper for checking uniqueNameKind (#9121)
* Factor helper for checking uniqueNameKind

Makes some code more concise.
I've opted not to use this in all places that I could, because there are
some places that are doing checks for multiple unique name kinds, and
those are more efficient the way they're currently implemented.

* Use another helper while we're here

I noticed that we have a helper for this, so I figured I'd tidy up some
code.