Don't need to sort packages anymore (#9138)
* Don't need to sort packages anymore

This was only here because the `lower_bound` / `upper_bound` logic in
`EnforcePackagePrefix` depended on it.

Now that I've removed that, we can also delete the fast sort. This is
one less thing that we need to have for parity with all other symbols
(because symbol IDs don't change and thus we can't sort).

* Move lexCmp to Exports (not used for packages anymore)