Add a usage of `Root::NOT_IN_PACKAGE`
Right now, this doesn't report an error saying that `Root` is not
imported, because Sorbet thinks the owner of `Root::NOT_IN_PACKAGE` is
actually `Root::Nested` (the current package, where the usage is).

This is because Sorbet uses location information for where the symbol
was defined to figure out which package owns this symbol.

This is kind of a weird case, because it's a question of how Sorbet
recovers from reporting errors.

One thing that Sorbet could choose to do when recovering from the
"defined somewhere it wasn't supposed to be defined" is to... forcibly
set the canonical loc of `Root::NOT_IN_PACKAGE` to some file that's
owned by the `Root` package? Which would then make Sorbet report that
`Root` is not imported in this package, so `NOT_IN_PACKAGE` cannot be
used.

This is a case that will change in behavior once we stop using canonical
locs to determine package ownership, so I wanted to capture it.