Add `using namespace core::packages;`
This was required to get things like `FullyQualifiedName` and `Export`
to resolve after they got moved. I chose to add this instead of fully
qualifying those references because it's less typing and realistically
we shouldn't be defining new constants in here that would have shadowed
one of those names.

Then, having added it, I figured I may as well remove all the things
that are now redundant.