Compute package ownership using module nesting (#9160)
* Add packageRegistryOwner and package fields

* Replace some usages of getPackageNameForFile

* Track constantAssignmentDefinitions to skip errors for definitions

If we're okay reporting somewhat weird "bad visibility" errors for
constant literals that show up in definition positions (like the LHS of
a constant assignment), we could drop this logic. But I figured that we
don't want the errors to be confusing.

* Delete setPackageLocs

* Remove a packager hack in resolver

* One change in the test

Previously, Sorbet would think that `::GlobalBar` was owned by
`::<PackageSpecRegistry>::RBI`, because it was defined in a file owned
by that package.

After this change, that's no longer the case.

We'll have to chat about this and see what the expected behavior should
be here.

* Fix bug with A::B::C::D

* Ban package-private methods outside packages

* stog   //test/cli:update_autocorrect-private \

* tools/scripts/format_website.sh

* Remove ownsSymbol

* Don't need to deal with behavior anymore

* Inline a function

Consistency with the other places this new field is used

* Add a comment

* Add a comment

* Back out redundant change

* Replace packageForSymbol with ClassOrModule::package