Remove `PackageInfoImpl` (#9188)
* Deduplicate VisibleTo struct

* PackageInfoImpl into PackageInfo.h

* Remove unused struct

* Move the comments

* Replace PackageInfo with PackageInfoImpl

* Remove references to PackageInfoImpl

* Use std:: prefix a lot

* Actually need these non-pure virtual methods

* Move most big implementations to .cc file

* format

* s/core::packages:://

* Add missing header file

* Remove NonePackage

* Missing an include

* Forgot to put this in the namespace

* Remove some unused includes

* Add `using namespace core::packages;`

This was required to get things like `FullyQualifiedName` and `Export`
to resolve after they got moved. I chose to add this instead of fully
qualifying those references because it's less typing and realistically
we shouldn't be defining new constants in here that would have shadowed
one of those names.

Then, having added it, I figured I may as well remove all the things
that are now redundant.