minitest: Don't deepCopy a literal (#9198)
* minitest: Don't deepCopy a literal

For `it` blocks, we were creating an `InsSeq` and keeping around the
`Literal` node for the argument to the `it` block. But we don't even
keep around the `it` method call itself.

We only need to keep around the `it` argument if we're trying to support
hover / typechecking in the argument itself. But if the argument is just
a string literal, we don't need to keep that around. Let's omit it to
save memory.

* Update exp files