wip: Actually this abstract! alloc trick could work
the main difference between last time and this time was that I wasn't actually trying to keep around the ":boom: instantiated abstract class" exception, I was just trying to approximate it with print statements saying whether the slow method was called at all

but faithfully implementing the exception forced me to do the mod = self trick to put the abstract class in the closure. I had been trying to do it by defining new on the Hooks module, but if you do that, then all you have access to is self and that will point to Child for Child.new

zanker:   For the actual version, you’d maybe want to not call alias_method on each inheritance if the hook was unset to deal with heavy hierarchy classes since iirc alias_method can blow some method caches?

the old approach I had been trying to get working was something like this:
this breaks with your A/B/C/D test case but it even breaks with just the abstract! > abstract! test case
but it only puts the new method in the inheritance tree once, so there is only ever one super call. vs with the A/B/C/D case, we see two log lines for "called slow wrapper"

I think that to fix ^, you'd actually have to do something like making the second abstract! call actually re-define the wrapper method. What happens is that when abstract! defines the wrapper method with extend(Hooks), that only runs once—if Hooks is already in the hierarchy, it doesn't run a second time. So you'd need something like "for second and subsequent calls to abstract!, you need to do the define_singleton_method approach or something

ok yeah, something like this works too

^ this has the benefit that there are no block-defined methods, there's only method-methods. Not sure whether that matters or not for performance. And it still has the problem that you get multiple copies of the "called slow wrapper" method for deep hierarchies, but those shouldn't matter because in that case it's going to raise anyways