minitest: Fix `it(...) { describe { ... } }` (#9290)
* no-op: Remove bodyIsClass

I was grepping for `isClass` and it was not showing me the two most
relevant locations in the file, so I wanted to change the code to show
them.

* no-op: recurse -> tryRunSingleOnSend

* no-op: clang-format

* no-op: Co-locate runSingle, tryRunSingleOnSend, prepareBody

You basically always have to read `tryRunSingleOnSend` and `prepareBody`
at the same time. `runSingle` is a little bit more of a standalone
method, but they're still all tightly coupled.

* Add test showing current behavior

* Delete some code to fix the test

I don't really understand the original motivation of this code.

But deleting it seems to fix the issue I'm seeing, and it doesn't seem
to break anything else.

I'm going to test on Stripe's codebase and make sure.