minitest: Also handle anonymous `it { ... }` tests (#9291)
* no-op: Use new `nameForTestHelperMethod` helper

* no-op: Combine the `after`, `before`, and `it` cases

The `RuntimeMethodDefinition` comment is the only difference here
(because the `nameForTestHelperMethod` already handles the `before` and
`after` methods, see #9288).

But we actually want that logic to apply, so that it gets the same
treatment as `it`-defined methods do. See #9199 for a discussion of that
change, and how it drops memory usage.

* Show behavior before

* minitest: Also handle anonymous `it { ... }` tests

Minitest supports this out of the box, so Sorbet may as well, too.