Put it back to a single ClassDef
This avoids adding another loc to `Sorbet::Private::Static`.
That probably doesn't matter in practice, but it was not a change I was
intending to make and the only reason I made it was to avoid a little
typing, so I may as well keep things the way they were.