Speed up untracking package references
On the slow path, the package references sets will start off empty, so
looping over them should be fast. However, because calls to untrack and
track were interleaved, the untrack calls would encounter the entries
added by the track calls and do extra work that's not needed (since the
track calls would a file A, but the untrack call would be for some other
file B). By doing all the untracking upfront, we can make it faster.