Delete `NumBlock`
The only different between a regular `Block` and a `NumBlock` is that the parameters are modelled by a `parser::NumParams` node instead of an `Args` node (which should properly be called parser::Params`).

There's no need to have a distinct block type for it, and it simplies a lot of branching if we remove it.