prework: Add unsetAllExportedInPackage
By moving from packager.cc to VisibilityChecker.cc, duplicate export
checking will now run on the fast path.

The new algorithm, instead of being powered by sorting a deduplicating
`FullyQualifiedName` objects for the exports, is going to be based on
whether a symbol is **already** exported by the time an `export` line
marks it as exported.

That means that we have to start from a starting point where nothing in
the package is exported, so here's a helper that does that.