no-op: Isolate base case and recursive case
Better ignoring whitespace.

The old version of `recursiveSetIsExported` unfolded the recursive case
one level. That was ~fine before, when logic that needs to be done at
each level is trivial (e.g., `flags.isExported = true`).

But I plan to add more logic here, and I didn't want to have to
duplicate it in both the base and recursive cases.