Add `RequiresAncestorKernel` rewriter (#9345)
By default Sorbet treats `Module`s as they could be mixed into any
class, which includes classes that don't descend from `Kernel`, i.e.
subclasses of `BasicObject`. While this is the correct way to treat
modules in general, and is safe by default, this treatment makes it
really difficult to work with modules in Ruby. In almost every case, the
user intent is to mix a module into a class that descends from `Kernel`,
and only seldomly is that expectation not true.

This commit makes that default expectation explicit when
`requires_ancestor` feature is enabled. This is done by rewriting
modules to include a `requires_ancestor { Kernel }` call, so that
modules can refer to `Kernel` methods by default. If a module is meant
for being mixed into a class that doesn't descend from `Kernel` the
escape hatch is to mark it explicitly as `requires_ancestor {
BasicObject }` in which case, this rewriter doesn't do anything.

Update rewriter/RequiresAncestorKernel.cc

Co-authored-by: Alexander Momchilov <amomchilov@users.noreply.github.com>