Use `<Magic>.requires_ancestor`
It is **never** okay to insert an `include` or an `extend` in a rewriter
pass to get a Sorbet-specific helper method like
`mixes_in_class_methods`, `abstract!`, `requires_ancestor`, etc.
defined.

Doing so masks valid errors: e.g. had someone went to define `abstract!`
in this class, they would be met with runtime-only missing methods
errors, because `extend T::Helpers` is not in the ancestor hierarchy at
runtime.